Refactor: Organize app into /app directory and update auth flow
