+++ app/(auth)/login/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useState } from 'react';
-import { useRouter } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { signInWithEmail, signInWithGoogle } from '@/lib/firebaseAuth';
 import Link from 'next/link';
 
@@ -10,6 +10,8 @@ export default function Login() {
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
   const router = useRouter();
+  const params = useSearchParams();
+  const redirect = params.get('redirect') || '/app';
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -18,7 +20,7 @@ export default function Login() {
 
     try {
       await signInWithEmail(email, password);
-      router.push('/');
+      router.push(redirect);
     } catch (error: unknown) {
       const errorMessage = error instanceof Error ? error.message : 'An error occurred';
       setError(errorMessage);
@@ -32,7 +34,7 @@ export default function Login() {
       setLoading(true);
       setError('');
       await signInWithGoogle();
-      router.push('/');
+      router.push(redirect);
     } catch (error: unknown) {
       const errorMessage = error instanceof Error ? error.message : 'An error occurred';
       setError(errorMessage);
@@ -42,10 +44,9 @@ export default function Login() {
   };
 
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50">
-      <div className="max-w-md w-full space-y-8 p-8 bg-white rounded-lg shadow">
-        <h2 className="text-2xl font-bold text-center">Login to AuraX</h2>
-        
+    <div className="min-h-screen flex items-center justify-center">
+      <div className="max-w-md w-full space-y-8 p-8 glass rounded-2xl">
+        <h2 className="text-2xl font-bold text-center">Welcome back to Aura</h2>
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
             <input
@@ -53,28 +54,27 @@ export default function Login() {
               value={email}
               onChange={(e) => setEmail(e.target.value)}
               required
-              className="w-full px-3 py-2 border rounded-md"
+              className="w-full px-3 py-2 rounded-md bg-transparent border border-white/20"
               placeholder="Email"
             />
           </div>
-          
           <div>
             <input
               type="password"
               value={password}
               onChange={(e) => setPassword(e.target.value)}
               required
-              className="w-full px-3 py-2 border rounded-md"
+              className="w-full px-3 py-2 rounded-md bg-transparent border border-white/20"
               placeholder="Password"
             />
           </div>
 
-          {error && <div className="text-red-500 text-sm">{error}</div>}
+          {error && <div className="text-red-400 text-sm">{error}</div>}
 
           <button
             type="submit"
             disabled={loading}
-            className="w-full py-2 px-4 bg-blue-500 text-white rounded-md disabled:opacity-50"
+            className="w-full py-2 px-4 rounded-xl bg-gradient-to-r from-neon-purple via-neon-pink to-neon-cyan text-black font-semibold disabled:opacity-50"
           >
             {loading ? 'Signing in...' : 'Sign In'}
           </button>
@@ -83,20 +83,20 @@ export default function Login() {
         <button
           onClick={handleGoogleLogin}
           disabled={loading}
-          className="w-full py-2 px-4 bg-white border border-gray-300 rounded-md disabled:opacity-50"
+          className="w-full py-2 px-4 rounded-xl border border-white/20 hover:bg-white/10 disabled:opacity-50"
         >
-          Sign in with Google
+          Continue with Google
         </button>
 
         <div className="text-center text-sm space-y-2">
           <p>
             Don&apos;t have an account?{' '}
-            <Link href="/signup" className="text-blue-500">
+            <Link href={`/signup?redirect=${encodeURIComponent(redirect)}`} className="underline">
               Sign up
             </Link>
           </p>
           <p>
-            <Link href="/forgot-password" className="text-blue-500">
+            <Link href="/forgot-password" className="underline">
               Forgot your password?
             </Link>
           </p>
@@ -105,3 +105,4 @@ export default function Login() {
     </div>
   );
 }
+
