+++ app/(auth)/signup/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useState } from 'react';
-import { useRouter } from 'next/navigation';
+import { useRouter, useSearchParams } from 'next/navigation';
 import { signUpWithEmail } from '@/lib/firebaseAuth';
 import Link from 'next/link';
 
@@ -10,6 +10,8 @@ export default function Signup() {
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
   const router = useRouter();
+  const params = useSearchParams();
+  const redirect = params.get('redirect') || '/app';
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -18,7 +20,7 @@ export default function Signup() {
 
     try {
       await signUpWithEmail(email, password);
-      router.push('/');
+      router.push(redirect);
     } catch (error: unknown) {
       const errorMessage = error instanceof Error ? error.message : 'An error occurred';
       setError(errorMessage);
@@ -28,10 +30,9 @@ export default function Signup() {
   };
 
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50">
-      <div className="max-w-md w-full space-y-8 p-8 bg-white rounded-lg shadow">
+    <div className="min-h-screen flex items-center justify-center">
+      <div className="max-w-md w-full space-y-8 p-8 glass rounded-2xl">
         <h2 className="text-2xl font-bold text-center">Create Account</h2>
-        
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
             <input
@@ -39,28 +40,27 @@ export default function Signup() {
               value={email}
               onChange={(e) => setEmail(e.target.value)}
               required
-              className="w-full px-3 py-2 border rounded-md"
+              className="w-full px-3 py-2 rounded-md bg-transparent border border-white/20"
               placeholder="Email"
             />
           </div>
-          
           <div>
             <input
               type="password"
               value={password}
               onChange={(e) => setPassword(e.target.value)}
               required
-              className="w-full px-3 py-2 border rounded-md"
+              className="w-full px-3 py-2 rounded-md bg-transparent border border-white/20"
               placeholder="Password (min. 6 characters)"
             />
           </div>
 
-          {error && <div className="text-red-500 text-sm">{error}</div>}
+          {error && <div className="text-red-400 text-sm">{error}</div>}
 
           <button
             type="submit"
             disabled={loading}
-            className="w-full py-2 px-4 bg-blue-500 text-white rounded-md disabled:opacity-50"
+            className="w-full py-2 px-4 rounded-xl bg-gradient-to-r from-neon-purple via-neon-pink to-neon-cyan text-black font-semibold disabled:opacity-50"
           >
             {loading ? 'Creating account...' : 'Create Account'}
           </button>
@@ -68,7 +68,7 @@ export default function Signup() {
 
         <p className="text-center text-sm">
           Already have an account?{' '}
-          <Link href="/login" className="text-blue-500">
+          <Link href={`/login?redirect=${encodeURIComponent(redirect)}`} className="underline">
             Login
           </Link>
         </p>
@@ -76,3 +76,4 @@ export default function Signup() {
     </div>
   );
 }
+
