+++ app/app/boosts/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useEffect, useMemo, useState } from 'react';
 import { db, auth } from '@/lib/firebase';
-import { addDoc, collection, doc, increment, onSnapshot, serverTimestamp, setDoc, updateDoc } from 'firebase/firestore';
+import { addDoc, collection, doc, increment, serverTimestamp, setDoc } from 'firebase/firestore';
 import confetti from 'canvas-confetti';
 import { motion } from 'framer-motion';
 
@@ -22,7 +22,6 @@ const defaultBoosts: Boost[] = [
 
 export const dynamic = 'force-dynamic';
 
-// Legacy route kept for backward compatibility; redirect to new /app path
 export default function BoostsPage() {
   const [completing, setCompleting] = useState<string | null>(null);
 
@@ -58,9 +57,29 @@ export default function BoostsPage() {
     }
   }
 
-  if (typeof window !== 'undefined') {
-    window.location.replace('/app/boosts');
-  }
-  return null;
+  return (
+    <main>
+      <h1 className="mb-4 text-xl font-semibold">Aura Boosts</h1>
+      <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:grid-cols-3">
+        {defaultBoosts.map((b) => (
+          <motion.div
+            key={b.id}
+            className="glass rounded-2xl p-4 hover:shadow-glow"
+            whileHover={{ y: -2 }}
+          >
+            <div className="mb-2 text-sm uppercase tracking-wider text-[rgba(230,230,255,0.75)]">{b.type}</div>
+            <div className="mb-4 text-lg font-medium">{b.title}</div>
+            <button
+              onClick={() => complete(b)}
+              disabled={completing === b.id}
+              className="w-full rounded-xl bg-gradient-to-r from-neon-purple via-neon-pink to-neon-cyan px-4 py-2 font-semibold text-black disabled:opacity-60"
+            >
+              {completing === b.id ? 'Completing...' : `Complete (+${b.points})`}
+            </button>
+          </motion.div>
+        ))}
+      </div>
+    </main>
+  );
 }
